# docs/openapi.yaml
openapi: 3.0.0
info:
  title: CinePal API
  description: Backend API for CinePal movie booking system. Handles movie catalogues, theatre management, seat bookings, and payments.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    ClerkAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Clerk JWT token. Use `Bearer <token>`.
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
paths:
...
  /bookings:
    get:
      summary: List all bookings (Supports Search & Pagination)
      security:
        - ClerkAuth: []
      parameters:
        - name: q
          in: query
          description: Search by User ID or Booking ID
          schema: {type: string}
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 10}
      responses:
        200:
          description: Paginated list of bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Atomic seat hold
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                showtimeId: {type: string}
                seats:
                  type: array
                  items:
                    type: object
                    properties:
                      row: {type: string}
                      col: {type: integer}
      responses:
        201:
          description: Pending booking created
        409:
          description: Seats unavailable
  /bookings/mine:
    get:
      summary: List my bookings
      security:
        - ClerkAuth: []
      responses:
        200:
          description: User bookings
  /bookings/{id}:
    get:
      summary: Get booking details
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        200:
          description: Booking details
    put:
      summary: Swap seats for pending booking
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seats:
                  type: array
                  items:
                    type: object
                    properties:
                      row: {type: string}
                      col: {type: integer}
      responses:
        200:
          description: Seats swapped
    delete:
      summary: Cancel pending booking
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        200:
          description: Cancelled
  /bookings/{id}/cancel:
    put:
      summary: Cancel confirmed booking
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        200:
          description: Cancelled with potential refund
...